requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
xgboost
mlflow
joblib
pyarrow
pytest
pyyaml